backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai==1.51.0
langchain==0.3.1
langchain-openai==0.2.1
langchain-community==0.3.1
langchain-chroma==0.1.4
chromadb==0.5.11
gitpython==3.1.43
tiktoken==0.8.0
pydantic==2.9.2
httpx==0.27.2
